requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
grpcio==1.70.0
grpcio-tools==1.68.1
idna==3.10
protobuf==5.29.0
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.3.0
